+++ src/hamster/edit_activity.py
@@ -50,6 +50,7 @@ class CustomFactController(Controller):
         self._gui = load_ui_file("edit_activity.ui")
         self.window = self.get_widget('custom_fact_window')
         self.window.set_size_request(600, 200)
+        self.window.set_default_size(1000, 200)
 
 
         self.action = action
@@ -81,6 +82,9 @@ class CustomFactController(Controller):
 
         self.tags_entry = widgets.TagsEntry(parent=self.get_widget("tags box"))
 
+        self.exported_checkbox = gtk.CheckButton(label=_("do not export"))
+        self.get_widget("exported box").add(self.exported_checkbox)
+
         self.save_button = self.get_widget("save_button")
 
         # this will set self.master_is_cmdline
@@ -94,7 +98,8 @@ class CustomFactController(Controller):
         elif action == "clone":
             base_fact = runtime.storage.get_fact(fact_id)
             self.fact = base_fact.copy(start_time=dt.datetime.now(),
-                                       end_time=None)
+                                       end_time=None,
+                                       exported=False)
         else:
             self.fact = Fact(start_time=dt.datetime.now())
 
@@ -122,6 +127,7 @@ class CustomFactController(Controller):
         self.activity_entry.connect("changed", self.on_activity_changed)
         self.category_entry.connect("changed", self.on_category_changed)
         self.tags_entry.connect("changed", self.on_tags_changed)
+        self.exported_checkbox.connect("toggled", self.on_exported_toggled)
 
         self._gui.connect_signals(self)
         self.validate_fields()
@@ -276,6 +282,11 @@ class CustomFactController(Controller):
             self.fact.tags = self.tags_entry.get_tags()
             self.update_cmdline()
 
+    def on_exported_toggled(self, widget):
+        if not self.master_is_cmdline:
+            self.fact.exported = self.exported_checkbox.get_active()
+            self.update_cmdline()
+
     def present(self):
         self.window.present()
 
@@ -301,6 +312,7 @@ class CustomFactController(Controller):
         self.category_entry.set_text(self.fact.category)
         self.description_buffer.set_text(self.fact.description)
         self.tags_entry.set_tags(self.fact.tags)
+        self.exported_checkbox.set_active(self.fact.exported)
         self.validate_fields()
 
     def update_status(self, status, markup):